# ggml llama: align structs for memory optimization on 64-bit platforms