--- a/examples/perplexity/perplexity.cpp
+++ b/examples/perplexity/perplexity.cpp
@@ -691,13 +691,13 @@ static bool decode_helper(llama_context * ctx, llama_batch & batch, std::vector<
         const int32_t n_tokens = std::min(n_batch, (int32_t) (batch.n_tokens - i));
 
         llama_batch batch_view = {
-            n_tokens,
             batch.token    + i,
             nullptr,
             batch.pos      + i,
             batch.n_seq_id + i,
             batch.seq_id   + i,
             batch.logits   + i,
+            n_tokens,
             0, 0, 0, // unused
         };