@slaren 2024-05-14
The changes to the llama.h public structs are effectively a API breaking change for no real benefit. The other structs are less sensitive since they are internal to ggml, but I don't see how this is worth the risk of introducing bugs.