@USBhost 2024-05-14
What I find interesting "sample time" has regressed about 0.02 ms for me. Not like that has much meaning but still interesting.
```
Master:
llama_print_timings:      sample time =       7.35 ms /   100 runs   (    0.07 ms per token, 13607.29 tokens per second)
llama_print_timings:      sample time =       7.29 ms /   100 runs   (    0.07 ms per token, 13711.78 tokens per second)
llama_print_timings:      sample time =       7.30 ms /   100 runs   (    0.07 ms per token, 13704.26 tokens per second)
llama_print_timings:      sample time =       7.59 ms /   100 runs   (    0.08 ms per token, 13173.49 tokens per second)
llama_print_timings:      sample time =       7.24 ms /   100 runs   (    0.07 ms per token, 13821.70 tokens per second)

PR:
llama_print_timings:      sample time =       8.61 ms /   100 runs   (    0.09 ms per token, 11611.70 tokens per second)
llama_print_timings:      sample time =       8.49 ms /   100 runs   (    0.08 ms per token, 11773.02 tokens per second)
llama_print_timings:      sample time =       8.69 ms /   100 runs   (    0.09 ms per token, 11514.10 tokens per second)
llama_print_timings:      sample time =       8.59 ms /   100 runs   (    0.09 ms per token, 11646.87 tokens per second)
llama_print_timings:      sample time =       8.65 ms /   100 runs   (    0.09 ms per token, 11566.04 tokens per second)
llama_print_timings:      sample time =       8.58 ms /   100 runs   (    0.09 ms per token, 11650.94 tokens per second) 
```