@GermanAizek 2024-05-14
@USBhost, Yes, it is possible. This is due to fact that C++ compiler does not always know whether optimization is necessary in certain places or not, in this way we strictly inform compiler that we need certain sizes classes/structures.