@GermanAizek 2024-05-14
> What I find interesting "sample time" has regressed about 0.02 ms for me. Not like that has much meaning but still interesting.

Which compiler did you compile on for which platform?