@USBhost 2024-05-14
> > What I find interesting "sample time" has regressed about 0.02 ms for me. Not like that has much meaning but still interesting.
> 
> Which compiler did you compile on for which platform?
```
         .://:`              `://:.            usbhost@IONA
       `hMMMMMMd/          /dMMMMMMh`          ------------
        `sMMMMMMMd:      :mMMMMMMMs`           OS: Proxmox VE 8.2.2 x86_64
`-/+oo+/:`.yMMMMMMMh-  -hMMMMMMMy.`:/+oo+/-`   Host: Super Server 0123456789
`:oooooooo/`-hMMMMMMMyyMMMMMMMh-`/oooooooo:`   Kernel: 6.8.4-3-pve
  `/oooooooo:`:mMMMMMMMMMMMMm:`:oooooooo/`     Uptime: 5 hours, 33 mins
    ./ooooooo+- +NMMMMMMMMN+ -+ooooooo/.       Packages: 1133 (dpkg)
      .+ooooooo+-`oNMMMMNo`-+ooooooo+.         Shell: bash 5.2.15
        -+ooooooo/.`sMMs`./ooooooo+-           Resolution: 1024x768
          :oooooooo/`..`/oooooooo:             Terminal: /dev/pts/0
          :oooooooo/`..`/oooooooo:             CPU: AMD EPYC 7F72 (48) @ 3.200GHz
        -+ooooooo/.`sMMs`./ooooooo+-           GPU: NVIDIA RTX A6000
      .+ooooooo+-`oNMMMMNo`-+ooooooo+.         GPU: NVIDIA RTX A4000
    ./ooooooo+- +NMMMMMMMMN+ -+ooooooo/.       GPU: NVIDIA RTX A4000
  `/oooooooo:`:mMMMMMMMMMMMMm:`:oooooooo/`     Memory: 2538MiB / 257591MiB
`:oooooooo/`-hMMMMMMMyyMMMMMMMh-`/oooooooo:`
`-/+oo+/:`.yMMMMMMMh-  -hMMMMMMMy.`:/+oo+/-`
        `sMMMMMMMm:      :dMMMMMMMs`
       `hMMMMMMd/          /dMMMMMMh`
         `://:`              `://:`
```
main: built with cc (Debian 12.2.0-14) 12.2.0 for x86_64-linux-gnu

`./llama-bench -m /mnt/36TB/AI/llama-3-8B-Instruct-More-abliterated/ggml-model-f16.gguf -t 24 -r 10 -pg 512,128

Master:
| model                          |       size |     params | backend    |    threads |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ------------: | ---------------: |
| llama 8B F16                   |  14.96 GiB |     8.03 B | CPU        |         24 |         pp512 |     57.80 ± 0.25 |
| llama 8B F16                   |  14.96 GiB |     8.03 B | CPU        |         24 |         tg128 |      8.70 ± 0.02 |
| llama 8B F16                   |  14.96 GiB |     8.03 B | CPU        |         24 |   pp512+tg128 |     27.02 ± 0.10 |

PR:
| model                          |       size |     params | backend    |    threads |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ------------: | ---------------: |
| llama 8B F16                   |  14.96 GiB |     8.03 B | CPU        |         24 |         pp512 |     58.99 ± 0.64 |
| llama 8B F16                   |  14.96 GiB |     8.03 B | CPU        |         24 |         tg128 |      8.72 ± 0.01 |
| llama 8B F16                   |  14.96 GiB |     8.03 B | CPU        |         24 |   pp512+tg128 |     27.07 ± 0.08 |`

I'm currently retrying at -r20