@GermanAizek 2024-05-14
@USBhost, I have:

```
$ lsb_release -a
No LSB modules are available.
Distributor ID: Debian
Description:    Debian GNU/Linux 12 (bookworm)
Release:        12
Codename:       bookworm
```

```
$ uname -a
Linux debian-laptop 6.1.0-21-amd64 #1 SMP PREEMPT_DYNAMIC Debian 6.1.90-1 (2024-05-03) x86_64 GNU/Linux
```

```
$ free -h
               total        used        free      shared  buff/cache   available
Mem:           6.7Gi       3.2Gi       1.2Gi        15Mi       2.6Gi       3.5Gi
Swap:          5.0Gi       2.4Gi       2.6Gi
```

```
$ cat /proc/cpuinfo 
processor       : 0
vendor_id       : AuthenticAMD
cpu family      : 23
model           : 17
model name      : AMD Ryzen 5 2500U with Radeon Vega Mobile Gfx
stepping        : 0
microcode       : 0x810100b
cpu MHz         : 1437.121
cache size      : 512 KB
physical id     : 0
siblings        : 8
core id         : 0
cpu cores       : 4
apicid          : 0
initial apicid  : 0
fpu             : yes
fpu_exception   : yes
cpuid level     : 13
wp              : yes
flags           : fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush mmx fxsr sse sse2 ht syscall nx mmxext fxsr_opt pdpe1gb rdtscp lm constant_tsc rep_good nopl nonstop_tsc cpuid extd_apicid aperfmperf rapl pni pclmulqdq monitor ssse3 fma cx16 sse4_1 sse4_2 movbe popcnt aes xsave avx f16c rdrand lahf_lm cmp_legacy svm extapic cr8_legacy abm sse4a misalignsse 3dnowprefetch osvw skinit wdt tce topoext perfctr_core perfctr_nb bpext perfctr_llc mwaitx cpb hw_pstate ssbd ibpb vmmcall fsgsbase bmi1 avx2 smep bmi2 rdseed adx smap clflushopt sha_ni xsaveopt xsavec xgetbv1 clzero irperf xsaveerptr arat npt lbrv svm_lock nrip_save tsc_scale vmcb_clean flushbyasid decodeassists pausefilter pfthreshold avic v_vmsave_vmload vgif overflow_recov succor smca sev sev_es
bugs            : sysret_ss_attrs null_seg spectre_v1 spectre_v2 spec_store_bypass retbleed smt_rsb srso div0
bogomips        : 3992.26
TLB size        : 2560 4K pages
clflush size    : 64
cache_alignment : 64
address sizes   : 43 bits physical, 48 bits virtual
power management: ts ttp tm hwpstate eff_freq_ro [13] [14]
```

```
$ glxinfo -B
name of display: :0
display: :0  screen: 0
direct rendering: Yes
Extended renderer info (GLX_MESA_query_renderer):
    Vendor: AMD (0x1002)
    Device: AMD Radeon Vega 8 Graphics (raven, LLVM 15.0.6, DRM 3.49, 6.1.0-21-amd64) (0x15dd)
    Version: 22.3.6
    Accelerated: yes
    Video memory: 1024MB
    Unified memory: no
    Preferred profile: core (0x1)
    Max core profile version: 4.6
    Max compat profile version: 4.6
    Max GLES1 profile version: 1.1
    Max GLES[23] profile version: 3.2
Memory info (GL_ATI_meminfo):
    VBO free memory - total: 524 MB, largest block: 524 MB
    VBO free aux. memory - total: 3340 MB, largest block: 3340 MB
    Texture free memory - total: 524 MB, largest block: 524 MB
    Texture free aux. memory - total: 3340 MB, largest block: 3340 MB
    Renderbuffer free memory - total: 524 MB, largest block: 524 MB
    Renderbuffer free aux. memory - total: 3340 MB, largest block: 3340 MB
Memory info (GL_NVX_gpu_memory_info):
    Dedicated video memory: 1024 MB
    Total available memory: 4438 MB
    Currently available dedicated video memory: 524 MB
OpenGL vendor string: AMD
OpenGL renderer string: AMD Radeon Vega 8 Graphics (raven, LLVM 15.0.6, DRM 3.49, 6.1.0-21-amd64)
OpenGL core profile version string: 4.6 (Core Profile) Mesa 22.3.6
OpenGL core profile shading language version string: 4.60
OpenGL core profile context flags: (none)
OpenGL core profile profile mask: core profile

OpenGL version string: 4.6 (Compatibility Profile) Mesa 22.3.6
OpenGL shading language version string: 4.60
OpenGL context flags: (none)
OpenGL profile mask: compatibility profile

OpenGL ES profile version string: OpenGL ES 3.2 Mesa 22.3.6
OpenGL ES profile shading language version string: OpenGL ES GLSL ES 3.20
```

master:
```
main: n_len = 32, n_ctx = 2048, n_kv_req = 32

<s> Hello my name is Danielle. I am 21 years old. I am a full time student at a local community college. I am majoring

main: decoded 27 tokens in 877.91 s, speed: 0.03 t/s

llama_print_timings:        load time =   67847.49 ms
llama_print_timings:      sample time =       0.99 ms /    28 runs   (    0.04 ms per token, 28311.43 tokens per second)
llama_print_timings: prompt eval time =   32733.53 ms /     5 tokens ( 6546.71 ms per token,     0.15 tokens per second)
llama_print_timings:        eval time =  877862.89 ms /    27 runs   (32513.44 ms per token,     0.03 tokens per second)
llama_print_timings:       total time =  945758.49 ms /    32 tokens

```

PR:
```
main: n_len = 32, n_ctx = 2048, n_kv_req = 32

<s> Hello my name is Danielle. I am 21 years old. I am a full time student at a local community college. I am majoring

main: decoded 27 tokens in 878.38 s, speed: 0.03 t/s

llama_print_timings:        load time =   67512.84 ms
llama_print_timings:      sample time =       0.97 ms /    28 runs   (    0.03 ms per token, 28747.43 tokens per second)
llama_print_timings: prompt eval time =   32895.30 ms /     5 tokens ( 6579.06 ms per token,     0.15 tokens per second)
llama_print_timings:        eval time =  878320.06 ms /    27 runs   (32530.37 ms per token,     0.03 tokens per second)
llama_print_timings:       total time =  945892.14 ms /    32 tokens
```